Fixed a regression introduced by revision a30d7ca57e
The problem occurred on anchor links like "something.html#fragment". name_index would then be 0 because the filename is at the start of the path, so atrribute_value.at( name_index - 1 ) would go out of bounds.

A check for name_index == 0 fixes the problem and makes sure that we still prevent matches for suffixes of other filenames.