Fixed a rare crash issue when loading epub files.
Qt classes like QHash and QList are *instance* reentrant, not *member function* reentrant like the Qt docs say. It's funny, the Qt docs for both of these classes say "Note: All functions in this class are reentrant.", but other parts of Qt docs say "By extension, a class is said to be reentrant if its member functions can be called safely from multiple threads, as long as each thread uses a different instance of the class. The class is thread-safe if its member functions can be called safely from multiple threads, even if all the threads use the same instance of the class."

This makes the "All functions in this class are reentrant" statement misleading, since serialization needs to happen at the instance level, not function level.

So when ImportOEBPS::LoadOneFile accessed the ImportOEBPS instance hashes and lists in a multithreaded way, it could cause memory corruption.

Tracking this down took a while, *especially* since it's impossible to reproduce deterministically.