Fixed a rare issue with no tab opened by default when loading epubs.
The problem appears when you re-load an epub, and the only tab you had open was the first HTML file in the reading order. Then the BB sends a command to the TabManager to open the first HTML file in the new epub, and that new file has the same filename as the old tab that is just being closed. The TabManager thinks that no new tab needs to be opened since a tab with the file of the same filename is already opened.