Fixed a rare crash/memory corruption issue with automatic OPF updates
For TextResource::SetText, we need to delay updating the QTextDocument
if it has been called from something other than the main GUI thread. Why?
Because a CodeView is probably connected to the text document,
and if we update it from a non-GUI thread, it will notify the
CodeView base class to update as well and that will crash us since
the base class derives from QWidget (and those can only be updated
in the GUI thread).
  So we cache the text update into m_Cache and update the QTextDocument
when we return to the GUI thread. The single-shot timer makes sure
of that.